fix(directions): rename mapPage and convert userLocation var to useState
